% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tvem.R
\name{print.tvem}
\alias{print.tvem}
\title{print.tvem:  Print output from a model that was fit by the tvem function.}
\usage{
\method{print}{tvem}(x, ornate = TRUE, ...)
}
\arguments{
\item{x}{The tvem object (output of the tvem or select_tvem function)}

\item{ornate}{Whether to print lines between different sections of the output for easier reading.}

\item{...}{Further arguments currently not supported}
}
\description{
print.tvem:  Print output from a model that was fit by the tvem function.
}
